+++ packages/plugin-image/web/src/editor.ts
@@ -5,7 +5,7 @@ import { DEFAULTS } from './consts';
 import { EditorPluginCreator } from 'wix-rich-content-common';
 import { createImageData } from './createImageData';
 import { createRicosExtensions } from './tiptap';
-import { TiptapEditorPlugin } from 'wix-tiptap-editor';
+import { TiptapEditorPlugin } from 'ricos-tiptap-types';
 
 export const pluginImage: EditorPluginCreator<ImagePluginEditorConfig> = config => {
   const pluginConfig: ImagePluginEditorConfig = { ...DEFAULTS.config, ...config };