+++ packages/ricos-schema/web/src/generate-plugin-defaults.ts
@@ -16,6 +16,7 @@ import {
   Node_Type,
   Decoration_Type,
   LinkData,
+  ColorData,
   MentionData,
 } from '../generated/stringEnums/wix/rich_content/v1/';
 
@@ -127,6 +128,7 @@ const generateDecorationDefaults = () =>
       [Decoration_Type.LINK]: () => ({
         linkData: LinkData.fromJSON({ link: { url: '', anchor: '', rel: {} } }),
       }),
+      [Decoration_Type.COLOR]: () => ColorData.fromJSON({}),
       [Decoration_Type.MENTION]: () => ({
         mentionData: MentionData.fromJSON({ name: '', slug: '' }),
       }),