+++ packages/tiptap-editor/web/src/context.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import { RicosTiptapContextValue } from './models/extension-types';
+import { RicosTiptapContextValue } from 'ricos-tiptap-types';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const EditorPropsContext = React.createContext<Record<string, any> | null>(null);