+++ packages/tiptap-editor/web/src/extensions/extension-unique-id/index.ts
@@ -1,1 +1,0 @@
-export { createUniqueId } from './unique-id';