+++ CHANGELOG.md
@@ -20,13 +20,70 @@
 ### :bug: Bug Fix
 - `headings`
   - [#3163](https://github.com/wix/ricos/pull/3163) fix lists wiring to `documentStyle`
+- `tiptap-editor`
+  - [#3234](https://github.com/wix/ricos/pull/3234) core commands added to core configs
 
 </details>
 <hr/>
 
+## 8.65.5 (Dec 09, 2021)
+### :bug: Bug Fix
+
+- `toolbars-new/editorCommands`
+  - [#3028](https://github.com/wix/ricos/pull/3028) fix code-block action in the new toolbar
+  - [#3224](https://github.com/wix/ricos/pull/3224) fix font size display when empty blocks are in selection
+- `gallery`
+  - [#3225](https://github.com/wix/ricos/pull/3225) fix image margin on mobile sliders
+  - [#3228](https://github.com/wix/ricos/pull/3228) fix organize media settings top buttons push items down
+- `tooltip`  
+  - [#3218](https://github.com/wix/ricos/pull/3218) wire font family for BM theme
+
+### :house: Internal
+
+- `tiptap-editor`  
+  - [#3217](https://github.com/wix/ricos/pull/3217) pinned tiptap versions
+- `tiptap-extensions`  
+  - [#3229](https://github.com/wix/ricos/pull/3229) core extensions moved to separate package
+- `ricos-tiptap-types`  
+  - [#3229](https://github.com/wix/ricos/pull/3229) public tiptap related types moved to separate package
+- `toolbars-new`
+  - [#3232](https://github.com/wix/ricos/pull/3232) `toolbarInputButton` supports disable prop
+- `ui-components`
+  - [#3233](https://github.com/wix/ricos/pull/3233) refactor settings-mobile-header/button components
+
+## 8.65.4 (Dec 06, 2021)
+### :bug: Bug Fix
+
+- `ricos-content`
+  - [#3215](https://github.com/wix/ricos/pull/3215) fixed memory leak in fromDraft caused by generateId
+
+## 8.65.3 (Dec 06, 2021)
+
+### :bug: Bug Fix
+
+- `ricos-editor`
+  - [#3232](https://github.com/wix/ricos/pull/3232) fix new formatting toolbar crash
+
+## 8.65.2 (Dec 05, 2021)
+
+### :bug: Bug Fix
+
+- `toolbars-new`
+  - [#3211](https://github.com/wix/ricos/pull/3211) font size dropdown a11y
+
+## 8.65.1 (Dec 05, 2021)
+
+### :bug: Bug Fix
+
+- `tiptap-editor`
+  - [#3210](https://github.com/wix/ricos/pull/3210) set focus to editor in decoration commands
+- `tiptap-editor`
+  - [#3206](https://github.com/wix/ricos/pull/3206) add mentions plugin
+
 ## 8.65.0 (Dec 05, 2021)
 
 ### :rocket: New Feature
+
 - `*`
   - [#3203](https://github.com/wix/ricos/pull/3203) icons bm theme wiring
   - [#3207](https://github.com/wix/ricos/pull/3207) bg & hover color bm theme wiring
@@ -76,6 +133,7 @@
   - [#3196](https://github.com/wix/ricos/pull/3196) new dropdown style and a11y fixes for safari
 
 ## 8.63.7 (Nov 30, 2021)
+
 ### :bug: Bug Fix
 
 - `button`